requirements.txt
allure-pytest==2.15.2
allure-python-commons==2.15.2
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
exceptiongroup==1.3.1
execnet==2.1.2
greenlet==3.2.4
idna==3.11
iniconfig==2.1.0
packaging==25.0
playwright==1.57.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.2
pytest==8.4.2
pytest-base-url==2.1.0
pytest-dependency==0.6.0
pytest-order==1.3.0
pytest-playwright==0.7.1
pytest-xdist==3.8.0
python-dotenv==1.2.1
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.32.5
soupsieve==2.8.3
text-unidecode==1.3
tomli==2.3.0
typing_extensions==4.15.0
urllib3==2.6.2